accessible/tests/mochitest/name/test_tree.xhtml

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="Accessibility Name Calculating Test.">

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />

  <script type="application/javascript"
          src="../treeview.js" />

  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../role.js"></script>
  <script type="application/javascript"
          src="../name.js"></script>
  <script type="application/javascript"
          src="../events.js"></script>

  <script type="application/javascript">
  <![CDATA[
    async function doTest() {
      let treeNode = getNode("tree");
      let reordered = waitForEvent(EVENT_REORDER, treeNode);
      treeNode.view = new nsTreeTreeView();
      await reordered;

      const tree = {
        role: ROLE_OUTLINE,
        children: [
          {
            role: ROLE_LIST
          },
          {
            role: ROLE_OUTLINEITEM,
            children: [],
            name: "row1col"
          },
          {
            role: ROLE_OUTLINEITEM,
            children: [],
            name: "row2_col"
          },
          {
            role: ROLE_OUTLINEITEM,
            children: [],
            name: "row2.1_col"
          },
          {
            role: ROLE_OUTLINEITEM,
            children: [],
            name: "row2.2_col"
          },
          {
            role: ROLE_OUTLINEITEM,
            children: [],
            name: "row3_col"
          },
          {
            role: ROLE_OUTLINEITEM,
            children: [],
            name: "row4col"
          }
        ]
      };
      testAccessibleTree(treeNode, tree);

      let tableNode = getNode("table");
      reordered = waitForEvent(EVENT_REORDER, tableNode);
      tableNode.view = new nsTableTreeView(2);
      await reordered;

      const tableTree = {
        role: ROLE_TABLE,
        children: [
          {
            role: ROLE_LIST
          },
          {
            role: ROLE_ROW,
            children: [
              {
                role: ROLE_GRID_CELL,
                children: [],
                name: "row0_t_col1"
              },
              {
                role: ROLE_GRID_CELL,
                children: [],
                name: "row0_t_col2"
              }
            ],
            name: "row0_t_col1 row0_t_col2"
          },
          {
            role: ROLE_ROW,
            children: [
              {
                role: ROLE_GRID_CELL,
                children: [],
                name: "row1_t_col1"
              },
              {
                role: ROLE_GRID_CELL,
                children: [],
                name: "row1_t_col2"
              }
            ],
            name: "row1_t_col1 row1_t_col2"
          }
        ]
      };
      testAccessibleTree(tableNode, tableTree);

      let treetableNode = getNode("treetable");
      reordered = waitForEvent(EVENT_REORDER, treetableNode);
      treetableNode.view = new nsTableTreeView(2);
      await reordered;

      const treetableTree = {
        role: ROLE_TREE_TABLE,
        children: [
          {
            role: ROLE_LIST
          },
          {
            role: ROLE_ROW,
            children: [
              {
                role: ROLE_GRID_CELL,
                children: [],
                name: "row0_tt_col1"
              },
              {
                role: ROLE_GRID_CELL,
                children: [],
                name: "row0_tt_col2"
              }
            ],
            name: "row0_tt_col1 row0_tt_col2"
          },
          {
            role: ROLE_ROW,
            children: [
              {
                role: ROLE_GRID_CELL,
                children: [],
                name: "row1_tt_col1"
              },
              {
                role: ROLE_GRID_CELL,
                children: [],
                name: "row1_tt_col2"
              }
            ],
            name: "row1_tt_col1 row1_tt_col2"
          }
        ]
      };
      testAccessibleTree(treetableNode, treetableTree);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  ]]>
  </script>

  <hbox flex="1" style="overflow: auto;">

  <body xmlns="http://www.w3.org/1999/xhtml">
    <a target="_blank"
       href="https://bugzilla.mozilla.org/show_bug.cgi?id=546812"
       title="Treegrid row accessible shouldn't inherit name from tree accessible">
      Mozilla Bug 546812
    </a>
    <a target="_blank"
       href="https://bugzilla.mozilla.org/show_bug.cgi?id=664376"
       title="Table rows of XUL trees no longer containing cell content as accessible name">
      Mozilla Bug 664376
    </a>
    <p id="display"></p>
    <div id="content" style="display: none">
    </div>
    <pre id="test">
    </pre>
  </body>

  <vbox flex="1">

  <tree id="tree" flex="1">
    <treecols>
      <treecol id="col" flex="1" primary="true" label="column"/>
    </treecols>
    <treechildren/>
  </tree>

  <tree id="table" flex="1">
    <treecols>
      <treecol id="t_col1" flex="1" label="column"/>
      <treecol id="t_col2" flex="1" label="column 2"/>
    </treecols>
    <treechildren/>
  </tree>

  <tree id="treetable" flex="1">
    <treecols>
      <treecol id="tt_col1" flex="1" label="column" primary="true"/>
      <treecol id="tt_col2" flex="1" label="column 2"/>
    </treecols>
    <treechildren/>
  </tree>

  </vbox> <!-- close tests area -->
  </hbox> <!-- close main area -->
</window>
